Fixed to follow clang-tidy

+++ localization/landmark_based_localizer/ar_tag_based_localizer/src/ar_tag_based_localizer_core.cpp
@@ -219,7 +219,7 @@ void ArTagBasedLocalizer::image_callback(const sensor_msgs::msg::Image::ConstSha
     image_pub_.publish(out_msg.toImageMsg());
   }
 
-  const int detected_tags = markers.size();
+  const int detected_tags = static_cast<int>(markers.size());
 
   diagnostic_msgs::msg::DiagnosticStatus diag_status;
